Gui: Fix undo/redo behaviour in transformation tool
If the user clicks on undo while the transformation dialog is open the currently active transaction will be closed.
From now on any change of the placement won't be recorded any more so that e.g. canceling the dialog or an undo after
accepting the dialog leads to unexpected behaviour.

To fix the issue two new virtual methods onUndo() and onRedo() are added to TaskDialog and reimplemented in TaskCSysDragger.
These functions make sure to open a new transaction.

This fixes issue 19152