Gui: Fix linking error on Linux by not defining _USE_3DCONNEXION_SDK
For Linux builds the define _USE_3DCONNEXION_SDK is actually not needed. Instead SPNAV_FOUND is used.
PR 17000 has caused a regression by always setting the define _USE_3DCONNEXION_SDK which led to the linking error if the spacenav library is
not installed on the system because some source files won't be included into the build.

Solution:
Make sure that for systems other than APPLE or MSVC the define will never be set.