Sketch: Make option 'ESC can leave Sketch edit mode' working again
In older versions the parameter was read-in in ViewProviderSketch::setEdit()
but after several code refactoring steps all parameter handling was moved to
ParameterObserver.

Since for internal handling this class uses a different string than the
actual paramter name the observer fails to handle the parameter change.
Instead the parameter is only correctly handled when creating the sketch.

Solution:
To fix the problem the internal string should be equal to the parameter
name.

This fixes issue 18944