Gui: Fix command StdCmdProperties
The command is only able to show the property view in case it's invisible. But it fails to raise the
widget. And if in the preferences 'Combined' mode is set it fails completely.

Solution:
The method DockWindowManager::activate already does everything what's needed. Now the command only
must be changed to access the right docked widget.