Gui: Fix crash when trying to transform link object
This is a regression caused by PR 17564.

ViewProviderDragger has the member 'forwardedViewProvider' that is used to handle the editing by the parent object.
This means that inside ViewProviderLink::startEditing 'csysDragger' can be null but this isn't checked so that
accessing the member causes a segmentation fault or a failing assert in debug mode, respectively.

Solution:
Make sure that csysDragger is not null before accessing it.

This fixes 19542.