Part: Fix FCBRepAlgoAPI_BooleanOperation::Build
Since OCCT 7.6 the Build() method has an optional argument. When overriding this method in the subclass FCBRepAlgoAPI_BooleanOperation
the optional argument must be kept as otherwise it will break ABI compatibility and may lead to unexpected behaviour.

This also fixes compiler warnings if OCCT version is >= 7.6 that otherwise won't be suppressed by the NOLINT comment.